feat(wallet): account-rename
QR CODE refactor for reuse
