+++ packages/store/slices/wallets.ts
@@ -11,6 +11,7 @@ import { RootState } from "../store";
 export interface StoreWallet {
   index: number;
   address: string;
+  name: string;
   provider: "native" | "ledger" | "google";
   network: NetworkKind;
   networkId: string;
@@ -55,6 +56,7 @@ const walletsSlice = createSlice({
   },
   reducers: {
     addSelected: storeWalletsAdapter.setOne,
+    updateWallet: storeWalletsAdapter.upsertOne,
     removeSelected: storeWalletsAdapter.removeOne,
     resetAllWallets: storeWalletsAdapter.removeAll,
     setSelectedNativeWalletIndex: (state, action: PayloadAction<number>) => {
@@ -78,7 +80,7 @@ const tokensSlice = createSlice({
   name: "tokens",
   initialState: storeTokens.getInitialState(),
   reducers: {
-    updateOne: storeTokens.upsertOne,
+    updateToken: storeTokens.upsertOne,
     removeToken: storeTokens.removeOne,
     resetAllTokens: storeTokens.removeAll,
   },
@@ -103,6 +105,10 @@ export const addressBookReducer = addressBookSlice.reducer;
 export const tokensReducer = tokensSlice.reducer;
 export const { addEntry, removeEntry, resetAllAddressBook } =
   addressBookSlice.actions;
-export const { addSelected, resetAllWallets, setSelectedNativeWalletIndex } =
-  walletsSlice.actions;
-export const { updateOne, resetAllTokens } = tokensSlice.actions;
+export const {
+  addSelected,
+  resetAllWallets,
+  setSelectedNativeWalletIndex,
+  updateWallet,
+} = walletsSlice.actions;
+export const { updateToken, resetAllTokens } = tokensSlice.actions;
